Health_Backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
pydantic[email]
pydantic-settings